Some posts with reusable blocks are missing styles
Posts with reusable blocks that trigger either of the conditionals of `cycle_through_blocks()` that also have other blocks _before_ the reusable block will be missing some or all of the custom styles entered for other blocks on that post.

Returning nothing in these instances resets the `$style` variable, so returning the existing contents of `$style` seems to fix the issue.